libs/data-access/pyproject.toml
[tool.poetry]
name = "data-access"
version = "0.1.0"
description = ""
authors = ["PJ <pieter-jan@storychief.io>"]
readme = "README.md"
packages = [{ include = "data_access" }]

[tool.poetry.dependencies]
python = "^3.11"
gcloud = "^0.18.3"
google-cloud-bigquery = "^3.4.2"
gcloud-aio-bigquery = "^6.3.0"
gcloud-aio-pubsub = "^6.0.1"
gcloud-aio-storage = "^8.2.0"
pydantic = {extras = ["dotenv"], version = "^2.0.0"}
google-cloud-firestore = "^2.11.1"
google-api-python-client = "^2.90.0"
tenacity = "^8.2.2"
cryptography = "^41.0.6"
mypy = "^1.8.0"
aiohttp = "^3.9.2"
google-cloud-pubsub = "^2.21.5"
pydantic-settings = "^2.6.1"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.11.1"
pytest = "^7.3.2"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
